Correspondent: GET, POST, UPDATE, DELETE

+++ NPaperless.DataAccess.Interfaces/IRepository.cs
@@ -3,7 +3,7 @@ namespace NPaperless.DataAccess.Sql;
 public interface IRepository<T>
 {
     T Add(T item);
-    T FindById(int id);
+    T FindById(long id);
     IEnumerable<T> GetAll();
     T Update(T item);
     void Delete(T item);
